% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichhelper-package.R
\docType{package}
\name{enrichhelper-package}
\alias{enrichhelper}
\alias{enrichhelper-package}
\title{enrichhelper: Helper Functions for Gene Set Enrichment Analysis}
\description{
Wrapper functions around 'clusterProfiler' for over-representation analysis (ORA) and gene set enrichment analysis (GSEA), using either a Bioconductor 'OrgDb' or a manually supplied term-to-gene map, plus a Cleveland dotplot visualization for enrichment results.
}
\author{
\strong{Maintainer}: Jelmer Poelstra \email{jelmerpoelstra@gmail.com}

Authors:
\itemize{
  \item Jelmer Poelstra \email{jelmerpoelstra@gmail.com}
}

}
\keyword{internal}
