% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{run_gsea}
\alias{run_gsea}
\title{Run a Gene Set Enrichment Analysis (GSEA)}
\usage{
run_gsea(
  df,
  contrast,
  term_map = NULL,
  OrgDb = NULL,
  keyType = "ENTREZID",
  ontology_type = NULL,
  GO_ontology = NULL,
  p_enrich = 0.05,
  allow_dups = FALSE,
  return_df = FALSE
)
}
\arguments{
\item{df}{Differential expression results; should at least have columns
\code{contrast}, \code{gene}, \code{lfc}/\code{log2FoldChange}.}

\item{contrast}{Contrast ID, should be one of the values in the
\code{contrast} column in \code{df}.}

\item{term_map}{Gene-to-term map (e.g., GO or KEGG). Either \code{term_map}
(use a 'manual' lookup table) or \code{OrgDb} (use a Bioconductor lookup
table) is required.}

\item{OrgDb}{Bioconductor \code{OrgDb} (alternative to \code{term_map} for available
organisms).}

\item{keyType}{OrgDb gene ID type. Only applies when using an \code{OrgDb}
instead of a \code{term_map}.}

\item{ontology_type}{\code{"GO"} (\code{enrichGO()}) or \code{"KEGG"} (\code{enrichKEGG()}).
Only applies when using an \code{OrgDb} instead of a \code{term_map}.}

\item{GO_ontology}{Only for GO analysis with an \code{OrgDb} \emph{and} when
\code{return_df == TRUE} (default is \code{"BP"} there). In all other cases, GO
analyses are run across all 3 ontologies (BP, CC, MF). (The reason this
applies for GO with an OrgDb while keeping the clusterProfiler format is
that clusterProfiler errors when trying to run \code{enrichGO()} with all
ontologies at once.)}

\item{p_enrich}{Adjusted p-value threshold for enrichment.}

\item{allow_dups}{Allow a gene ID to be present multiple times in a
(single-contrast) list of DEGs. This should typically \emph{not} be the
case, but could be so when working with gene IDs (orthologs) from
another species than the focal species. NOTE: this computes the mean
LFC across duplicated genes.}

\item{return_df}{Convert the result object to a simple dataframe
(tibble).}
}
\value{
A GSEA result object (\code{return_df = FALSE}) or a tibble
(\code{return_df = TRUE}).
}
\description{
A wrapper around \code{clusterProfiler::GSEA()}/\code{clusterProfiler::gseGO()} that
ranks genes by log2 fold change for a given contrast and runs the GSEA
algorithm against either a manually supplied term map or a Bioconductor
\code{OrgDb}.
}
